This patch changes the commit procedure (and thus jfsck()) to avoid writing the data we're about to commit into the transaction file.
This patch changes the commit procedure (and thus jfsck()) to avoid writing
the data we're about to commit into the transaction file.

The reason behind this is that it slows things down and doesn't provide any
real warantees: a crash is as likely to happen when we're writing the
transaction file as it is when we're writing the real file.

So by writing only the previous data (to be able to rollback it in the case of
a crash) we provide the same warantees, improve performance and reduce code.