This is an upgrade to the "tests" directory which contains files used for testing.
This is an upgrade to the "tests" directory which contains files used for
testing.

In this patch you'll find only the README file, but that's because the
previous tests directory contained a lot of small files and I didn't think it
was worth the effort to include them here. Besides, most of them were binaries
anyway.