I can't believe I didn't see it before. It's so obvious.
I can't believe I didn't see it before. It's so obvious.

Instead of reading the current contents and write the previous data to the
disk, we should write current data. The library doesn't have the "j" just
because.

While at it, resurrect the old J_NOROLLBACK from the dead, so people that
don't need to rollback transactions will pay the penalty.

This is also paving the road for lingering transactions.