Some filesystems can zero files if crash occur in special places. This can lead to transaction file corruption if the data (but not the header) is zeroed out, so we need some method for verification.
Some filesystems can zero files if crash occur in special places. This can
lead to transaction file corruption if the data (but not the header) is zeroed
out, so we need some method for verification.

This patch implements a simple but effective checksum on the transaction
files, that is appended at their end.

The algorithm is quite small, based on RFC 1071, and it's the one used for IP.