The constants are broken, increidbly things work for most things today by pure luck. They are in this state because at some point (way in the past) jfs and jtrans flags were independant, but this has not been the case for a long time.
The constants are broken, increidbly things work for most things today by pure
luck. They are in this state because at some point (way in the past) jfs and
jtrans flags were independant, but this has not been the case for a long time.

The patch unifies the constants and puts sane numbers to them.

Thanks to Shehjar Tikoo who noticed this and send a patch.