When linking an application that doesn't use Large File Support with the library (which does), several problems might arise from the data size differences.
When linking an application that doesn't use Large File Support with the
library (which does), several problems might arise from the data size
differences.

It's very important that everything uses the same interface. Since there's no
portable way to force the application we're linking with uses LFS, we just
give an error in case they don't. It might seem uglier, but it's a lot safer.