This patch implements a preloader library for the UNIX API.
This patch implements a preloader library for the UNIX API.

You can use it like:

LD_PRELOAD=libjio_preload.so my_nice_app

and it will intercept all the relevant calls to the UNIX API (like open(),
read(), write() and so on) and call libjio's instead.

It has some limitations: it works only with glibc, it has a fixed (build-time
configurable) maximum number of file descriptors (8k by default), and it
hasn't been tested under 64 bit platforms.

It also has some good points: it works with apps compiled both with and
without LFS, it doesn't introduce much overhead, and it allows you to use and
test the library without having to write or even recompile any applications.

The patch also includes some modifications to the Makefile.

In order to build it, do "make preload" and then "make preload_install" to
install it.