Implement journal relocation
This patch implements jmove_journal(), which is used to relocate the journal
directory. It takes a struct jfs and the new path, and does the change. It's
not atomic, so you have to take care of calling this without any other
operation touching the file. Because I expect most people to call this right
after jopen(), it's not much of a problem (you have to be really careful if
you relocate somewhere else).

It also updates the checker, the python bindings and the preloader library.

Finally, it breaks the API and ABI by adding a new field to struct jfs, and
changing jfsck() and jfsck_cleanup(). It sucks, but I rather do it now.

Signed-off-by: Alberto Bertogli <albertito@gmail.com>