Split the default CFLAGS in optional and mandatory parts
The library cannot be built without some flags, so split CFLAGS in two,
making it easier for the user to pass them in the command line without
having to include the mandatory ones.

While at it, also pass the flags when linking, because there may be
relevant flags. We should use LDFLAGS for this, tho; but for now we'll
take the simpler approach.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>