Use the appropriate data types in Python bindings
When PyArg_ParseTuple() expects an int, it's better if we really give it
an int. Otherwise, if we use a larger data type, we might get garbage
because it will assume it's the size of an int.

We had some cases when that could happen, and this patch gets rid of them.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>