Add support for libfiu, a library for fault injection
This patch implements support for libfiu, which is a library for fault
injection. It adds several "failure points" inside the code, that can (and
will) be used, in the future, for testing purposes.

In order to avoid a build-time dependency on libfiu, it also adds a
special libfiu local header to the tree.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>