jfsck(): Check there's enough room for the checksum before computing it
Before computing the checksum in jfsck(), see if the file is long enough
to contain it. Otherwise, if the transaction file was truncated, we
would check the checksum anyways.

This is usually not a problem because the checksums shouldn't match, but
it's better to play safe. Also, it properly report such transactions as
broken instead of corrupt.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>