Add behaviour tests
This patch adds a suite of tests that check the library's behaviour. It's
written in Python, so it uses the Python bindings.

It has three different "test suites", one of them uses the fault injection
support incorporated in a previous patch, so it depends on libfiu being
installed and libjio having been built with FI=1. The other two test
suites can be run without libfiu.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>