Call fdatasync() after writing to the real file
If O_SYNC was not passed to jopen(), then we could loose data because
jtrans_commit() assumes writes to the file are synchronous. That behaviour
was undocumented, so the user could have assumed O_SYNC was not necessary.

This patch implements a much more sane semantic by calling fdatasync() on
the file fd after writing the transaction when lingering transactions are
not enabled, avoiding the need to open the file with O_SYNC.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>