Use posix_fadvise() in jtrans_add()
When adding an operation to a transaction, it's highly likely that the
affected region will be read when committing.

This patch lets the kernel know that using posix_fadvise(), so it can take
actions based on that.

Because posix_fadvise() was included in SUSv3 and so far we've relied just
on SUSv2, we introduce a new header, compat.h, to provide a stub for that
function so the library can be built in systems that do not implement it.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>