[API BREAK] Use opaque types instead of structures in the public API
To make the API more tidy, simple and easier to use, this patch removes
the public struct jfs and struct jtrans, replacing them with opaque types
jfs_t and jtrans_t.

As a result, jopen() and jtrans_init() now return pointers to the
structures, and jclose() and jtrans_free() free them.

There is also a small but important semantic change in jtrans_commit(),
which now returns -2 on unrecovered errors (previously, it was necessary
for the caller to check the transaction flags).

The version number is also bumped to 0.50, to reflect this major API
change.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>