[API BREAK] Remove jfsck_cleanup(), make jfsck() cleanup by default
Since cleaning up is the normal behaviour, it makes no sense to keep a
separate function for doing it.

This patch removes jfsck_cleanup(), and makes jfsck() do it by default
instead. A flags parameter to jfsck() was added, and a J_NOCLEANUP flag
was added, to give the user the possibility of avoiding cleaning the
journal up.

In the future, the jfsck() code will probably get merged with
jfsck_cleanup(), but it's not done here to make the patch more robust.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>