[FORMAT CHANGE] Improve the on-disk transaction format
This patch changes the way transactions are stored on disk. The
fundamental reason for this change is to put the number of operations
contained in the transaction at the end.

That will allow us to write operations as they are added to the
transaction, instead of waiting until commit time, which can improve
performance on some operation-intensive workloads.

Other changes include the addition of a much needed version number (which
will hopefully keep the library backwards compatible in case of future
disk format changes), a more robust checksum algorithm, and defined
endianness of the stored data for portability purporses.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>