[API BREAK] Remove jfsck_return.apply_error
If jfsck() returns successfully, it should be possible to assume the file
is in a consistent state. However, currently if jfsck() fails to apply a
transaction it will report that in the jfsck_return structure, which is
misleading.

This patch changes that behaviour, making jfsck() return error in that
case, rendering the apply_error member useless.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>