jtrans_commit(): Do not unlink the journal file if rollback failed
When the commit fails we attempt to rollback. If that fails too, we should
not unlink the journal file, because then there is no chance jfsck()
can recover from the possible corruption.

This patch makes jtrans_commit() keep that file, assuming the caller will
know that it should stop working on it and attempt to run jfsck().

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>