libjio: Try harder to corrupt transaction files as a last resort
Overwriting the header is not enough, because if after doing so a new
transaction gets committed, we will first write the header and it's
possible that we crash right there, leaving an apparently OK transaction
file while it should actually be bogus.

This patch fixes that potential problem by overriding the trailer, so we
can be sure it will not be considered a valid transaction even after
partial reuse.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>