libjio: Mark the journal as broken when we should not continue committing
When we have very serious errors that can cause consistency problems if we
continue to operate on the file, mark the journal as broken by creating a
file named "broken". journal_new() will automatically fail if the file is
present, so no further transactions will be allowed.

Transactions that were being committed when we set the flag are not a
problem because they affect different portions of the file.

This protects from the following scenario:

 - Commit data X to a range A-B:
   - Write the transaction 1 to the journal
   - Write the data to the file
   - Attempt to remove the journal file: fail
 - Commit data Y to a range A-B: success
 - Crash
 - jfsck():
   - Finds transaction 1, writes data X to the file which corrupts things
     because data Y is the one that should be there.

We also return -2 in those cases, which means the user should abort
immediately, but this measure can help to prevent problems when the user
is not careful or can't stop further commits.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>