libjio: Make jfsck() remove the transactions and simplify jfsck_cleanup()
jfsck() is the one which should remove the transactions as it cleans them
up, and this patch makes it so. This went unnoticed for a long time
because it almost always performs a jfsck_cleanup() afterwards.

That also allow us to simplify jfsck_cleanup() by making it remove only
the files we know about, which currently is only the "lock" file, but
it could be more in the future. The "broken" file is explicitly ignored
because it should only be removed by jfsck().

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>