[API BREAK] Implement read operations
This patch implements read operations, allowing the users to read data
inside a transaction.

That means that transactions have now two kinds of operations inside them:
read and write.

They work in a similar way they do today: operations are applied in order,
when it's a write the data is written, when it's a read the data is read.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>