jfsck(): Lock the file before operating on it
Callers of jfsck() must not call it when the file is in use. However, it
doesn't hurt to protect from it by locking the file before using it, to
protect unwary users. The close() takes care of the unlock.

Note that if the process has locks held on the file, it's already in
trouble because it will loose them anyway as we're re-opening and closing
the file, so this patch does not change that situation.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>