bindings/python: Use ssize_t for s# parameters
Commit 9607741 (bindings/python: Add readv()/writev() support, 2009-07-13)
defined PY_SSIZE_T_CLEAN, which made PyArg_ParseTuple()'s s# parameters use
ssize_t for the string length, instead of int (which is the default). That was
needed for writev().

However, existing instances of s# were not converted, and caused subtle but
worrying memory corruption.

This patch fixes that by using ssize_t in the appropriate places.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>