tests/stress: Add a --force option
Sometimes it can be useful to force the tests to run, even if conflicting
options are selected.

Even if they're known to be broken, the way they break can be interesting when
developing.

This patch adds an --force option to jiostress to let the user do that in an
easy way.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>