python: Fix int/long argument parsing mismatch
Some PyArg_ParseTuple() invocations were told to parse to an int, but a long
was given to it instead.

On 64-bit platforms that mismatch caused trouble, but luckily nothing that
could cause corruption, since the resulting values were given to malloc(),
which would fail.

This issue came up when running the behaviour tests in Fedora 13.

The patch fixes this by telling PyArg_ParseTuple() we are giving it a long
instead of an int.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>