Fix Python setup.py so that it uses the "new" build directory
Python's setup.py was looking for the library (for build purposes) in the
libjio directory, instead of libjio/build where it should reside.

This broke the Python build, but went unnoticed because in all my repositories
I had (purely by chance) a shared library in the libjio directory.

This commit fixes this by making it look in the appropriate place, and also
making libjio/Makefile to create a link to the library with the soname, so the
Python interpreter can find the library (we don't need to install it because
ldconfig will take care of creating it).

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>