Make tests-fi run longer
To take advantage of the fault injection, we need to do more operations in the
stress test, otherwise we end up with only about 5 simulated failures.

This patch increases the number of operations from 50 to 400, which ends up
making about 40 simulated failures and taking about 2 minutes (overall, not
just the stress test) on my laptop (Thinkpad T400 with a P8400 CPU).

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>