bindings/python: Use more appropriate data types
In the bindings we often used long instead of either ssize_t or int, specially
for return values.

This patch changes the definitions to match the appropriate data type, and
adds checks for some cases where we need to do so in Python (e.g. we convert
to (signed) long long and use it as an (unsigned) off_t).

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>