tests: Add PYTHON3PATH hack to jiostress
If we want to use a custom python path when testing, it is inconvenience
because we use both python 2 and 3 in our tests, but both will be
looking at the same environment variable ($PYTHONPATH), and modules are
likely incompatible between the two.

This patch adds a hack to jiostress so we can run:

PYTHONPATH=xxx PYTHON3PATH=yyy make tests

and have things work the way we expect. It's not pretty, but it's
practical and not very intrusive.